[renderblocking] Register finish callbacks for link header preloads
This patch creates a non-null PendingLinkPreload object for each link
header preload, and make it owned by the Document object. This has two
effects:

1. Now the preloaded resource has an observer owned by the Document,
   so the preload won't be cancelled if we add and remove a client
   before the preload finishes. This fixes crbug.com/1311370.

2. Now render-blocking link headers can also use PendingLinkPreload
   to block and unblock rendering, sharing the same logic with link
   elements.

There is also change to WPT header-inserted-preload-link.tentative.html.
Since there's no element for a link header, the test case cannot listen
to the 'load' event of any EventTarget. Therefore, this patch also
modifies the test code to allow waiting for the PerformanceEntry for
the link header's url.

Bug: 1271296
Fixed: 1311370
Change-Id: Ic64d26341516a510079c01dbff0b0196c7ec9f24