Add anchor content attribute and hook up to popups [popup 6/N]
This CL adds support for the 'anchor' content attribute generally on
Element, and hooks it up to ancestor popup processing for popups.

This CL also enables the code that closes all popups when a <dialog>
is shown.

The test changes here are mostly cleanup and adding more cases, plus
fixing a few bugs that crept in to the copy/paste from <popup> element
tests.

Bug: 1307772
Change-Id: I3032d08ec30866290f10d2845b77f0c6e9037a76