Ensure fallback setup when manual slots are inserted into shadow DOM
When a new slot is added to a shadow DOM with manual slot assignment,
while the slot is guaranteed to not have any assigned node, it may still
have fallback content. This patch ensures that the fallback content is
correctly set up, so that it can be correctly rendered, and can be
correctly cleared when the slot gets an assignment.

Fixed: 1292292
Change-Id: I74403ab7c6003a9a0203cfdce1c2896f9529499e