ZoomForDSF: Add an optional fuzzy factor to layout table checks + use it
This patch adds an optional data-fuzzy-factor attribute to
check-layout-th.js. If set, then this provides an error tolerance to
whatever parameter is being checked.

If there is more than one parameter applied, this may cause confusion
as the fuzzy factor would apply to all of them. However, the alternative
is potentially doubling all of the properties that would permit
fuzziness

This patch also adds the fuzzy factor check to a test to make it pass.

R=ikilpatrick@chromium.org

Fixed: 1309972
Change-Id: I662c88a50a645fe9b35b0083d9f44774393096ed